shiftnotes-backend/requirements.txt
# Core Django (Latest LTS)
Django==4.2.7

# Django REST Framework
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9

# AWS
boto3==1.34.0

# Configuration  
python-decouple==3.8

# Development tools
django-extensions==3.2.3
